Removed use of CPP to guard Trustworthy
The CPP guard is only required for GHC 7.2 and older, but even Debian stable
packs GHC 7.4.  Also, CPP currently causes problems on OSX because of the
transition to clang so I'm currently being very conservative about using it.
When I originally started using the CPP guards GHC 6.12 was still a supported
compiler for `pipes`, but it's not any longer.